Continue work on atmospheric forcing verification.
Fix an error in the transformation of 0-360 CGRF longitudes to
NEMO -180/180.
That improves somewhat the agreement between calculated and forcing values.

Add note about Susan's investigation of the NEMO code that concluded that
the calculated wind components are subject to atmospheric stability criteria
manipulations and are rescaled to be in a neutral atmosphere,
so they do not directly reflect the applied forcing values.

Update relative file paths to reflect tools/I_ForcingFiles/ directory
re-organization.