Change recording of in-use NEMO-code & NEMO-forcing revisions.
Previously `hg heads` was being used but that records the repo tip revs,
not the present rev of the repo working directories which is misleading
if a repo has been updated to an earlier rev for debugging purposes.
Now, `hg parents` is used, and the file names in which the current
working directory revs are recorded have been changed to
`NEMO-code-rev.txt` and `NEMO-forcing_rev.txt`.