Relax restriction that all elements of list assignment be of same type.
That restriction is an attempt at Fortran namelist validation which is
out of scope for this module. Python lists in the resulting dict can
handle lists of heterogeneous values, and there are use cases for
heterogeneous lists in the NEMO code namelist. If type homogeneity in a
list is important in a particular application it should be handled in
the code using the results of the namelist parser, not by the parser
itself.