Make run_NEMO worker use today's date for nowcast & record it for forecasts.
This is to deal with the problem of slow west.cloud runs that push into
the day after the nowcast is run for.
Sending the nowcast run date to the manager causes it to be recorded in
the checklist so that it can be retrieved to set up the forecast and
forecast2 runs correctly,
and it can also be used to ensure that the download_results and
make_plot_files workers operate on the correct run results.