Fix Python 2 to 3 issue in nowcast.figures.plot_map() & delete plot_map_new().
The zip() builtin function in Python 3 returns an iterator object,
in contract to the list that was returned by  zip() in Python 2.
The matplotlib.patches.Polygon object constructor expects a list of
polygon x-y pairs, not an iterator.
So, the iterator returned by zip() is now explicitly transformed into
a list.