Add "station" dataset attribute from JSON locationCode
Modified onc_json_to_dataset() to include "station" as a dataset attribute from
the `parameters.locationCode` item in the JSON response. This is necessary
because "station" is the key that the `get_onc_ferry` worker expects to find an
element of the metadata it needs under. Updated unit tests to verify that the
"station" attribute exists and has the correct value after conversion.