Update `data_tools.onc_json_to_dataset()` to ONC API v3 JSON schema and add tests (#93)
* Update onc_json_to_dataset() to v3 API and add tests

Update the onc_json_to_dataset() function to handle ONC data web service API v3
response JSON schema. Also add unit tests for the function.

* Add "station" dataset attribute from JSON locationCode

Modified onc_json_to_dataset() to include "station" as a dataset attribute from
the `parameters.locationCode` item in the JSON response. This is necessary
because "station" is the key that the `get_onc_ferry` worker expects to find an
element of the metadata it needs under. Updated unit tests to verify that the
"station" attribute exists and has the correct value after conversion.