Improve geo_tools.find_closest_model_point() (#95)
* Fix failing find_closest_model_point() unit tests

Update unit tests that weren't adjusted when 'raiseOutOfBounds' parameter was
added to find_closest_model_point(). This ensures that a ValueError is raised
when raiseOutOfBounds is enabled and latitude and longitude values aren't on the
grid.

* Remove unused imports in geo_tools unit tests

The unnecessary import statements for netCDF4 and tidetools modules were removed
from the unit tests of the geo_tools module in the SalishSeaTools library. This
makes the unit tests leaner and reduces dependencies for running the tests.

* Add 'continental2.5' tolerances to find_closest_model_point()

Add tolerance values for HRDPS continental 2.5 km resolution rotated lat/lon
grid that became forcing for SalishSeaCast in Feb-2023.

* Code style gardening with black