Add Python 3.13 support (#124)
* Add Python 3.13 to GHA pytest-with-coverage matrix

Include Python 3.13 in the GitHub Actions workflow for pytest with coverage.
This ensures compatibility and testing for the latest Python version.

* Change to Python 3.13 for development

Update Python version to 3.13 in development and documentation environments, and
revise documentation to reflect Python 3.11 as the minimum supported version.
Bump various package versions across requirements and configuration files to
their latest stable releases.

* Update breaking changes doc for version 25.1

Document the incompatibility of version 25.1 with earlier versions, in
particular, the new minimum Python version requirement (3.11) and the use of
Python 3.13 for development.