Remove unnecessary dependencies from SalishSeaTools env files (#126)
* Drop unused 'cliff' dependency from env files

The 'cliff' package was removed from both environment-test.yaml and
environment-dev.yaml as it is no longer required because the command processor
packages that require it were removed from the repository long ago. This cleanup
helps streamline the environment configurations by reducing unnecessary
dependencies.

* Drop ipdb & jupyterlab from test environment yaml

This change streamlines the test environment by removing unnecessary
dependencies. ipdb and jupyterlab are not required for the testing workflow,
reducing setup complexity and potential overhead.