Docs build cleanup (#127)
* Drop SalishSeaNowcast sys.path entry in docs config

The sys.path entry for "../SalishSeaNowcast" was unnecessary and has been
removed. The SalishSeaNowcast package has been maintained in a separate
repository since late-2016.

* Remove unused deps from readthedocs env file

Jupyter and matplotlib have been removed from the environment-rtd.yaml file as
they are not directly required to build the documentation.

* Improve autodoc mock imports in docs configuration

Remove unused or redundant libraries from `autodoc_mock_imports` to streamline
the documentation build process. Added mocks for `matplotlib` and `numpy` that
were previously installed in the build environment. This reduces clutter and
improves the maintainability of the configuration file.