Handle retirement of water level stations at Halfmoon Bay and Squamish (#137)
* Change Halfmoon Bay tide gauge station to None

Set "stn number" for Halfmoon Bay to None to reflect the termination of
real-time water level observations there in Aug-2024. This ensures that
informative error messages are generated when water levels are requested at
Halfmoon Bay, and resolves the water level comparison figure generation failure
in SalishSeaNowcast.

* Change Squamish tide gauge station to None

Set "stn number" for Squamish to None to reflect the termination of
real-time water level observations there in Aug-2024. This ensures that
informative error messages are generated when water levels are requested at
Squamish, and resolves the water level comparison figure generation failure
in SalishSeaNowcast.