Split 2xrez comparision notebooks due to size
Moved the `plot_tile()` function to a separate module in preparation for having
to move to multiple comparison notebooks due to git complaining about the file
size having exceeded 10Mb.

Changed the `plot_tile()` function so that `vmax` is a parameter to enable
avoidance of normalization errors in tiles with no depths <15m (the default
value of `vmax` that was previously hard-coded).